Try to fix bug with filter assignment
The method _add_filter_number() was issuing redundant numbers when encountering multiple new, nameless filters.
I tried to fixed this and created a branch (fix_issue_151)

Also, note that if the filter has_a name, we do not assess whether or not it is new, ... is this OK??
I think so, because if it has_a name, it will be registered in ch_filter_dict which is returned by _xml_response_to_mt and then merged into the survey before the next channel is processed.

Relates to issue #151